Endre alle float til double. Laget condition i falling.cpp slik at ballen beveger seg mot mål-koord fra alle plasseringer. Bug: Når ballen kommer nær mål-koord, vil vinkelen variere mye pga unøyaktighet

+++ aconverter.h
@@ -2,20 +2,20 @@
 #define ACONVERTER_H
 #include <cmath>
 
-//constexpr float PI = 3.1415926;
+//constexpr double PI = 3.1415926;
 class  Aconverter{
 	public:
 	Aconverter() = default;
-	Aconverter(float a /*grader*/, float vel, int rel=200) : angle{a*(PI/180)}, velocity{vel}, relation{rel} {}
-	float x_velocity() {
+	Aconverter(double a /*grader*/, double vel, int rel=200) : angle{a*(PI/180.0)}, velocity{vel}, relation{rel} {}
+	double x_velocity() {
 		return velocity*cos(angle);
 	}
-	float y_velocity() {
+	double y_velocity() {
 		return velocity*sin(angle)*relation;
 	}
 	private:
-		float angle{0}; //Radianer
-		float velocity{0};
+		double angle{0}; //Radianer
+		double velocity{0};
 		int relation{200};//Relasjon mellom bevegelse i x- og y-aksen
 		
 };
